+++ include/resources/modelData.h
@@ -1,29 +1,13 @@
 // include/resources/modelData.h
 
 #pragma once
-#include <vector>
-#include "graphics/mesh.h"
-#include "assimp/Importer.hpp"
-#include "assimp/scene.h"
-#include "assimp/postprocess.h"
-
+#include "resources/materialData.h"
 namespace CubeDemo {
-using TexPtrArray = std::vector<TexturePtr>;
-using VertexArray = std::vector<Vertex>;
-using MeshArray = std::vector<Mesh>;
-
 
 // 包围球结构体
 struct BoundingSphere {
     vec3 Center; float Rad; // 包围球中心, 半径
     void Calc(const MeshArray& meshes); // 计算包围球
-};
-// 材质数据结构体
-struct MaterialData {
-    string Directory;
-    TexPtrArray LoadTex(aiMaterial* mat, aiTextureType type, const string& typeName);
-    void ProcMaterial(aiMesh* &mesh, const aiScene* &scene, TexPtrArray& textures);
-
 };
 
 // ModelData结构体
@@ -38,10 +22,11 @@ struct ModelData : public MaterialData {
     void LoadModel(const string& path); // 加载模型
 
 private:
+    mat4 m_ModelMatrix{mat4(1.0f)};
+    
     void ProcNode(aiNode* node, const aiScene* scene);
     Mesh ProcMesh(aiMesh* mesh, const aiScene* scene);
 
 };
-
 
 }   // namespace CubeDemo
