+++ include/utils/assimpKits.h
@@ -1,0 +1,6 @@
+// include/utils/assimpKits.h
+#pragma once
+
+#include "assimp/Importer.hpp"
+#include "assimp/scene.h"
+#include "assimp/postprocess.h"
